castをしやすいようにした

+++ DriveAction/Script/Scene/Play/Menu.h
@@ -55,7 +55,7 @@ public:
     void Draw()const;
 private:
     //メニュー画面を開いている時間の計測
-    float startTime;
+    double startTime;
     //メニューの背景
     int backScreen;
     //背景の明るさ
